+++ redis/src/lib.rs
@@ -43,7 +43,7 @@ use redis::{
 };
 
 /// A type alias for using `deadpool::Pool` with `redis`
-pub type Pool = deadpool::Pool<Connection, RedisError>;
+pub type Pool = deadpool::Pool<ConnectionWrapper, RedisError>;
 
 mod cmd_wrapper;
 pub use cmd_wrapper::{cmd, Cmd};
@@ -51,11 +51,16 @@ mod pipeline_wrapper;
 pub use pipeline_wrapper::{pipe, Pipeline};
 
 /// A type alias for using `deadpool::Object` with `redis`
-pub struct Connection {
+
+/// A wrapper for `redis::Connection`. The `query_async` and `execute_async`
+/// functions of `redis::Cmd` and `redis::Pipeline` consume the connection.
+/// This wrapper makes it possible to replace the internal connection after
+/// executing a query.
+pub struct ConnectionWrapper {
     conn: Option<RedisConnection>,
 }
 
-impl Connection {
+impl ConnectionWrapper {
     fn _take_conn(&mut self) -> RedisResult<RedisConnection> {
         if let Some(conn) = self.conn.take() {
             Ok(conn)
@@ -87,12 +92,12 @@ impl Manager {
 }
 
 #[async_trait]
-impl deadpool::Manager<Connection, RedisError> for Manager {
-    async fn create(&self) -> Result<Connection, RedisError> {
+impl deadpool::Manager<ConnectionWrapper, RedisError> for Manager {
+    async fn create(&self) -> Result<ConnectionWrapper, RedisError> {
         let conn = self.client.get_async_connection().compat().await?;
-        Ok(Connection { conn: Some(conn) })
+        Ok(ConnectionWrapper { conn: Some(conn) })
     }
-    async fn recycle(&self, conn: &mut Connection) -> Result<(), RedisError> {
+    async fn recycle(&self, conn: &mut ConnectionWrapper) -> Result<(), RedisError> {
         if conn.conn.is_some() {
             cmd("PING").execute(conn).await.map(|_| ())
         } else {